Fix `static_assert()` scope in `SizeClassAllocator32`.
It should be at the class scope and not inside the `Init(...)` function
because we want to error out as soon as the wrong type is constructed.
At the function scope the `static_assert` is only checked if the
function might be called.

This is a follow up to r349138.

rdar://problem/45284065

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@349959 91177308-0d34-0410-b5e6-96231b3b80d8