Relax test introduced in D65322
It is possible that addr2line returns a valid function and file name for
the passed address on some build configuations.
The test is only checking that asan_symbolize doesn't assert any more when
passed a valid file with an invalid address so there is no need to check
that it can't find a valid function name.
This should fix http://lab.llvm.org:8011/builders/sanitizer-x86_64-linux

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@370021 91177308-0d34-0410-b5e6-96231b3b80d8